### 2020-10-24 #2778 @whedon
[PRE REVIEW]: MiTfAT: A Python-based scikit-learn-friendly fMRI Analysis Tool, Made in Tuebingen.
**Submitting author:** @vahid-sb (<a href="http://orcid.org/0000-0001-6544-7960">Vahid S. Bokharaie</a>)
**Repository:** <a href="https://github.com/vahid-sb/MiTfAT" target ="_blank">https://github.com/vahid-sb/MiTfAT</a>
**Version:** v0.1.7
