+++ paper.bib
@@ -1,0 +1,12 @@
+@article{savic:2019,
+	title={{Early detection and monitoring of cerebral ischemia using calcium-responsive MRI probes}},
+	author={Savic, Tanja and Gambino, Giuseppe and Bokharaie, Vahid S and Noori, Hamid R and Logothetis, Nikos K and Angelovski, Goran},
+	journal={Proceedings of the National Academy of Sciences},
+	volume={116},
+	number={41},
+	pages={20666--20671},
+	year={2019},
+	publisher={National Acad Sciences}
+}
+
+
