### 2020-10-24 #2778 @vahid-sb
> Hi @vahid-sb, it looks like the header of your paper may not be valid—please take a look at the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography). It may just be missing the `---` line to close the YAML header section.
>
> Also, your reference(s) should be contained in a .bib file, rather than hardcoded in the paper.

HI @kyleniemeyer, both issues are resolved now. Thanks for drawing my attention to them. Seems I deleted '---' after the header. Didn't know it is part of .md file commands. Also, reference is now read from a bib file.
