### 2020-10-26 #2778 @arokem
👋 @JonathanReardon : Would you be willing to review this article for JOSS?
