### 2020-10-26 #2778 @arokem
@vahid-sb : while we are waiting for reviewers to come on-board, a few comments from me:

- There are several other software libraries that do things that are closely related to what MiTfAT does. For example, [nilearn](https://nilearn.github.io/). I think that you will need to add a section to your paper describing previous similar work and describing the relationship between the present library and this previous work.
- The word "Neuroscience" (first line) should not be capitalized.
- "belts and whistles" should be "bells and whistles". I should also add that I don't understand exactly what this sentence means. It seems to suggest that MiTfAT is somehow simpler to use than other software, but provides little evidence to support this assertion. I would rethink how you would like to say this in a manner that emphasizes design choices made by different libraries, rather than posing this as an assertion.
- Though novelty per-se is not a requirement, the text does imply that MiTfAT implements novel methods. What are these analysis methods?
- The word "accomponied" should be spelled "accompanied".
