### 2020-10-26 #2778 @JonathanReardon
> @JonathanReardon : thanks for your quick response here! We usually aim to finish all parts of the review within six weeks, so if you think that you could turn around comments in two-three weeks, that would be ideal.

Yes, absolutely fine! thanks
