+++ paper.md
@@ -20,7 +20,7 @@ bibliography: paper.bib
 
 # Summary
  
-Functional Magnetic Resonance Imaging, fMRI, is a technique used in Neuroscience to measure brain activity based on any signal that can be measured in an MRI scanner. Normally, fMRI is used to detect changes associated with blood flow. But it can also be used to detect changes in concentrations of molecules with different magnetic properties which are directly injected into the brain of a subject. 
+Functional Magnetic Resonance Imaging, fMRI, is a technique used in neuroscience to measure brain activity based on any signal that can be measured in an MRI scanner. Normally, fMRI is used to detect changes associated with blood flow. But it can also be used to detect changes in concentrations of molecules with different magnetic properties which are directly injected into the brain of a subject. 
 
 Regardless of the signal that is measured in fMRI recording, from a computational point of view, fMRI recordings will result in a number of time-series. And then those time-series should be analysed to find the answers to various questions of interest. The length of time-series depends on the number of time-steps in which we have measured the signals. And the number of these time-series depends on how-many voxels we have measured. A voxel is a 3-dimensional pixel or the unit of volume in which each fMRI signal is measured. What is the size of each of these voxels depends on the magnetic flux density of the MRI scanner, measures in Tesla (T). The higher is the magnetic flux density, the smaller are the voxels and the higher is the spatial resolution of the measurement. Hence, we end up with one time-series for each of the voxels arranged in a 3-dimensional structure. One characteristic of the fMRI measurements is that while they can have a high spatial resolution, and while they allow us to measure brain activity not only in the cortices, but also in deeper regions of the brain, but their temporal resolution is not normally high. Hence, we need the algorithms we use to analyse them might be very different than the ones we use to analyse time-series with higher temporal resolutions. And these algorithms can get complicated if we go beyond the very simple statistical analyses of the fMRI data.
 
@@ -28,9 +28,9 @@ Regardless of the signal that is measured in fMRI recording, from a computationa
 
 `MiTfAT` is a scikit-learn-friendly Python library to analyse fMRI data. It was primarily developed for the study which is presented in [@savic:2019].
 
-There are a few software packages that are commonly used by researchers to pre-process the fMRI time-series and then analyse them. But the decision to develop a new library which eventually led to `MiTfAT` library is motivated by a few reasons. The main reason is that all the commonly used fMRI analysis packages come with so many belts and whistles that it usually takes the user a long time to find out which parts of each software package is what he/she need. Another reason was that none of the commonly used software packages I could find was written in Python, which is my programming language of choice. And lastly, the analysis methods I needed for the problems I had in mind were not available in any of those software packages. 
+There are a few software packages that are commonly used by researchers to pre-process the fMRI time-series and then analyse them. But the decision to develop a new library which eventually led to `MiTfAT` library is motivated by a few reasons. The main reason is that all the commonly used fMRI analysis packages come with so many bells and whistles that it usually takes the user a long time to find out which parts of each software package is what he/she need. Another reason was that the kinds of analysis methods I needed for the problems I had in mind were not incorporated into those software packages. And lastly, most of the commonly used software packages I could find were not written in Python, which is my programming language of choice. A notable exception is `Nilearn` which includes various visualisation functionalities and machine learning tools to analyse fMRI data. But that library does not provide a comprehensive framework to contain various information and measurements related to of an experiment.  
 
[…]
