### 2020-10-26 #2778 @vahid-sb
> @vahid-sb : while we are waiting for reviewers to come on-board, a few comments from me:
>
>     * There are several other software libraries that do things that are closely related to what MiTfAT does. For example, [nilearn](https://nilearn.github.io/). I think that you will need to add a section to your paper describing previous similar work and describing the relationship between the present library and this previous work.
>
>     * The word "Neuroscience" (first line) should not be capitalized.
>
>     * "belts and whistles" should be "bells and whistles". I should also add that I don't understand exactly what this sentence means. It seems to suggest that MiTfAT is somehow simpler to use than other software, but provides little evidence to support this assertion. I would rethink how you would like to say this in a manner that emphasizes design choices made by different libraries, rather than posing this as an assertion.
>
>     * Though novelty per-se is not a requirement, the text does imply that MiTfAT implements novel methods. What are these analysis methods?
>
>     * The word "accomponied" should be spelled "accompanied".

Thanks, @arokem, and thanks @JonathanReardon for agreeing to review.
@arokem I have updated the paper correcting the typos you kindly reminded me of. About Nilearn, I am aware of that and it is one of the requirements for mitfat, although I have used it merely for IO operations. But mitfat treats the data in a more comprehensive manner, not simply as separate NumPy arrays that might or might not be related. I have defined a class for fMRI data which is the cornerstone of the code. I added a short description of that to the paper. And about novelty, it is true that what I was looking for, I could find in the common fMRI analysis packages, but that does not mean the methods themselves are new. Simply smart applications of undergraduate mathematics, that's how I would characterize them. For example, the detrending method discussed in Chapter 8 of the manual.
