### 2020-11-09 #2827 @whedon
[REVIEW]: MiTfAT: A Python-based scikit-learn-friendly fMRI Analysis Tool, Made in Tuebingen.
**Submitting author:** @vahid-sb (<a href="http://orcid.org/0000-0001-6544-7960">Vahid S. Bokharaie</a>)
**Repository:** <a href="https://github.com/vahid-sb/MiTfAT" target ="_blank">https://github.com/vahid-sb/MiTfAT</a>
**Version:** v0.1.7

- [x] **Installation:** Does installation proceed as outlined in the documentation? (https://github.com/vahid-sb/MiTfAT/issues/1)

- [x] **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is? (https://github.com/vahid-sb/MiTfAT/issues/3)
- [x] **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems). (https://github.com/vahid-sb/MiTfAT/issues/2)
