### 2020-11-25 #2827 @arokem
:wave: @JonathanReardon and @emdupre -- just checking: have you had a chance to take a look at the paper, and start checking off those items? If you post issues on the software repo, please do add a pointer to this issue, so they are cross-linked.
