### 2020-11-25 vahid-sb/MiTfAT#1 @emdupre
Package documentation: Downloading and running MiTfAT
Hi ! Thanks for writing this package and for submitting it to JOSS !

I tried to install MiTfAT and execute the initial example given in your documentation. Specifically, in a new conda environment, I ran `pip install mitfat` which succeeded.

Launching a new python terminal, I ran:

`import mitfat`.

Note that

```
import `MiTfAT`
```

as suggested in the documentation, fails. I then ran:

```
from mitfat.file_io import read_data
import pkg_resources

info_file = pkg_resources.resource_filename('mitfat', 'sample_info_file.txt')
DATA_PATH = pkg_resources.resource_filename('mitfat', 'datasets')
list_datasets = read_data(info_file)
```

which also fails with the following error.

```
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/emdupre/miniconda3/envs/MiTfAT/lib/python3.6/site-packages/mitfat/file_io.py", line 237, in read_data
    first_trial_time, trial_length = main_get_data(info_file_name)
  File "/home/emdupre/miniconda3/envs/MiTfAT/lib/python3.6/site-packages/mitfat/file_io.py", line 690, in main_get_data
    time_step = read_info_file(info_file_name)
  File "/home/emdupre/miniconda3/envs/MiTfAT/lib/python3.6/site-packages/mitfat/file_io.py", line 412, in read_info_file
    raise Exception('Where is the info file???')
```

I've confirmed in my file system that these files are not being downloaded with the library, likely because the `setup.py` script (specifically, [here](https://github.com/vahid-sb/MiTfAT/blob/a9dceaf022e815b047dd90da1378ba4512b36b4c/setup.py#L28)) suggests that these files are at the top level when they should be in the `tests` sub-directory.

Correctly packaging these files and updating the documentation appropriately would help in making sure this software is accessible !
