### 2020-11-29 vahid-sb/MiTfAT#1 @vahid-sb
Hi @emdupre. Thanks for pointing out the typo and the missing data files. Seems mistakenly I have removed a manifest.ini file which is needed to have the data files uploaded with the code. I'd update the code and the manual and will let you know about it when done.
