+++ setup.py
@@ -44,5 +44,6 @@ setup(
         "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
         "Programming Language :: Python",
         "Programming Language :: Python :: 3.7",
+	"Programming Language :: Python :: 3.8",
     ],
 )
