+++ README.rst
@@ -15,6 +15,8 @@ Introduction
 
 `MiTfAT` is a scikit-learn-friendly Python library to analyse fMRI data.
 
+Manual can be found `here<https://github.com/vahid-sb/MiTfAT/blob/master/docs/mitfat.pdf>`.
+
 There are a few software packages that are commonly used by researchers to pre-process the fMRI time-series and then analyse them. But the decision to develop a new library which eventually led to `MiTfAT` library is motivated by a few reasons. The main reason is that all the commonly used fMRI analysis packages come with so many belts and whistles that it usually takes the user a long time to find out which parts of each software package is what he/she need. Another reason was that none of the commonly used software packages I could find was written in Python, which is my programming language of choice. And lastly, the analysis methods I needed for the problems I had in mind were not available in any of those software packages.
 
 Hence the `MiTfAT` was developed. It is designed to be used for general fMRI time-series analysis, but in particular, signals obtained from molecular fMRI studies, i.e. the cases in which we measure the changes in concentration of molecules which might have been directly injected into the brain (which happens when the molecule is too big to pass through blood-brain-barrier).
