### 2020-12-04 vahid-sb/MiTfAT#1 @vahid-sb
Dear @emdupre, I updated the documentation. I added all the sample datasets as a zip file to repo and you can just download, unzip it and run the two scripts. They both should work if mitfat is installed. Please let me know how it goes.
