### 2020-12-17 vahid-sb/MiTfAT#1 @emdupre
Thanks, @vahid-sb ! I was able to install appropriately and access the `sample_config_file.txt` :+1:

I'm still not able to access the sample datasets directly in the package. For that, I think you'd need to edit your `setup.py` script.

You could change [this line](https://github.com/vahid-sb/MiTfAT/blob/06fa18950384079f438a0fd6cd1a59f25df3aca7/setup.py#L28) to something like:

```
package_data={
"mitfat": ["sample_info_file.txt"],
"tests": ["*"],
"tests.datasets": ["*"],
"tests.datasets_trials": ["*"]
},
```

But I'd actually encourage you to actually move this directory instead the `mitfat` directory and separate out `tests` (which could contain an inner `data` folder, for data used in testing) from `datasets` (which would contain the datasets used in your examples in the documentation). In which case, the above line could look like:

```
package_data={
"mitfat": ["sample_info_file.txt"],
"mitfat.tests.data": ["*.txt"],
"tests.datasets": ["*.nii.gz", "*.txt"],
"tests.datasets.dataset_trials": ["*.nii.gz", "*.txt"]
},
```

This would help to make sure that it's clear which data is used for internal library testing and which data is related to the documentation / user examples !
