### 2020-12-17 vahid-sb/MiTfAT#2 @emdupre
Package documentation: User examples
I was reviewing the new user documentation, and I'm having some trouble with sections 3.1 and 3.2 (in "Basics").

I assume that the `info_file` in the first 3.1 code block corresponds to `sample_config_file.txt`. But if I try to execute:

```
list_of_datasets = read_data(info_file)
```

I get the following error:

```
Config file: sample_config_file.txt
Current directory: /home/emdupre
Source folder you have given me:
 /home/emdupre/datasets
An exception has occurred, use %tb to see the full traceback.

SystemExit: Source folder DOES NOT exist! :/
```

I imagine the recommended path forward would be downloading `tests.zip` and placing it in the `datasets` folder, but it would be helpful if the documentation was explicit about this. Pending #1, I imagine even be inferred from the package directory, which would definitely help to quickly get new users running on sample data !

Alternatively, just reorganizing the documentation to explain the config file and which values should be changed before loading the data would be helpful for orienting new users.
