### 2020-12-17 vahid-sb/MiTfAT#3 @emdupre
Software paper: statement of need
Thank you for carefully considering how this software can help the fMRI research community ! I had a few comments on the statement of need that I think will help better situate this work.

First, on how it compares to other packages: I'm not sure that other software projects having more "bells and whistles" is an appropriate comparison point. Considering only Python neuroimaging analysis tools, [`fitlins`](https://fitlins.readthedocs.io/en/latest/), [`niworkflows`](https://github.com/nipreps/niworkflows), and [`NiBetaSeries`](https://joss.theoj.org/papers/10.21105/joss.01295) are all analysis packages focussing on very specific points of the (pre-) processing workflow. Similarly, even larger packages such as [`nilearn`](http://nilearn.github.io) can quickly orient new users through well-considered documentation.
Of course more could always be done, but I would encourage you to avoiding this positioning -- to my mind it's not a compelling argument for the need for MiTfAT.

Instead, I think it would make sense to focus specifically on those analyses that can be accomplished with MiTfAT that _cannot_ be accomplished with other software. In fact, I would position MiTfAT specifically as a package for fMRI with contrast agents. It seems that the majority of examples focus on this, and it is a use case that remains relatively unknown in most of the major Python analysis packages ! In line with this, I might spend more time explaining fMRI with contrast agents in the Summary, and then in the statement of need you can note that other packages do not consider this use case. This would also help to better differentiate this work from `nilearn`, with which it otherwise shares many similarities.

On references: I would encourage you to directly cite `nilearn` since you discuss it in text. It would also be helpful to include references on fMRI with contrast agents to orient the reader.

Another minor point: In the software statement of need, hierarchical clustering is described as a multi-stage process that first removes low SNR voxels before clustering.  But hierarchical clustering is more commonly used to refer to clustering algorithms such as [Ward](https://en.wikipedia.org/wiki/Ward%27s_method) or [hierarchical k-means](https://www.andrew.cmu.edu/user/hgifford/projects/k_means.pdf). Because this paragraph is potentially confusing to readers, I would encourage you to remove or refocus it.
