### 2021-01-11 #2827 @JonathanReardon
I have the software installed and all seems to be working as expected in terms of performance and functionality. In the README (first paragraph) it might help to be more explicit about what analysis methods you feel are missing from other software packages, and what software packages you are referring to. I would also re-frame the "belts and whistles" (*bells and whistles) sentence, it's quite vague and I think many would disagree that is the case. You could reword to suggest that other packages contain many options that can sometimes slow down the user (or something like that) -- just a suggestion.

From reading the README and manual, there appear to be quite a lot of grammatical errors and spelling mistakes, a final proofread would help. For example..

- "belts and whistles" should be "bells and whistles"
- "which parts of each software package is what he/she need" should be "which parts of each software package he/she needs"
- "Hence the MiTfAT was developed" should be "Hence MiTfAT was developed"

@vahid-sb
