+++ paper.md
@@ -26,11 +26,11 @@ Regardless of the signal that is measured in fMRI recording, from a computationa
 
 # Statement of need
 
-`MiTfAT` is a scikit-learn-friendly Python library to analyse fMRI data. It was primarily developed for the study which is presented in [@savic:2019].
+`MiTfAT` is a scikit-learn-friendly Python library to analyse fMRI data, with a focus on molecular fMRI experiments. It was primarily developed for the study which is presented in [@savic:2019].
 
-There are already a few Python packages that are used by researchers to pre-process the fMRI time-series and then analyse them, for example [`fitlins`](https://fitlins.readthedocs.io/en/latest/), [`niworkflows`](https://github.com/nipreps/niworkflows), and [`NiBetaSeries`](https://joss.theoj.org/papers/10.21105/joss.01295) which focus on very specific points of the analysis workflow. Or even the more comprehensive library [`nilearn`](http://nilearn.github.io) that includes various visualisation functionalities and machine learning tools to analyse fMRI data, but does not provide a ready-made framework to contain various information and measurements related to an experiment in molecular fMRI experiments. Hence the `MiTfAT` library was developed. It can be used for general fMRI time-series analysis, but in particular, signals obtained from molecular fMRI studies, i.e. the cases in which we measure the changes in concentration of molecules that might have been directly injected into the brain. The `MiTfAT` library incorporates all the information and data related to an experiment into a Python class object called `fmri_dataset`. And various attributes of this class can be used to identify all the data related to each experiment, and perform analyses on all. Such datasets can include various MRI measurements of the same subject, for example, T1-weighted and FISP signals, measured almost simultaneously. Or the dataset can include many trials in which the same set of stimuli is presented to a single subject repeatedly. 
+There are already a few Python packages that are used by researchers to pre-process the fMRI time-series and then analyse them, for example [`fitlins`](https://fitlins.readthedocs.io/en/latest/), [`niworkflows`](https://github.com/nipreps/niworkflows), and [`NiBetaSeries`](https://joss.theoj.org/papers/10.21105/joss.01295) which focus on very specific points of the analysis workflow. Or even the more comprehensive library [`nilearn`](http://nilearn.github.io) that includes various visualisation functionalities and machine learning tools to analyse fMRI data, but does not provide a ready-made framework to contain various information and measurements related to an experiment in molecular fMRI experiments. Hence the `MiTfAT` library was developed. It can be used for general fMRI time-series analysis, but in particular, signals obtained from molecular fMRI studies, i.e. the cases in which we measure the changes in concentration of molecules that might have been directly injected into the brain. The `MiTfAT` library incorporates all the information and data related to an experiment into a Python class object called `fmri_dataset`. And various attributes of this class can be used to identify all the data related to each experiment, and perform analyses on all. Such datasets can include various MRI measurements of the same subject, for example, T1-weighted and FISP signals measured almost simultaneously. Or the dataset can include many trials in which the same set of stimuli is presented or applied to a subject repeatedly. 
 
-The basic principle behind `MiTfAT` is that it imports all the relevant data of an fMRI experiment into a class object. The fMRI time-series are a member of this class and are stores as NumPy arrays. There are various functionalities available to analyse the data in a number of ways. They include:
+The basic principle behind `MiTfAT` is that it imports all the relevant data of an fMRI experiment into an object/class of type `fmri_dataset`. The fMRI time-series are a member of this class and are stored as NumPy arrays. There are various functionalities available to analyse the data in a number of ways. They include:
 
[…]
