+++ MANIFEST.in
@@ -1,0 +1,5 @@
+include requirements.txt
+include ./tests/*.*
+include ./tests/datasets/*.*
+include ./tests/dataset_trials/*.*
+
