+++ paper.bib
@@ -9,4 +9,28 @@
 	publisher={National Acad Sciences}
 }
 
+@article{kent:2019,
+  title={NiBetaSeries: task related correlations in fMRI},
+  author={Kent, James D and Herholz, Peer},
+  journal={Journal of Open Source Software},
+  volume={4},
+  number={41},
+  pages={1295},
+  year={2019}
+}
 
+@article{esteban:2019,
+  title={NiPreps: enabling the division of labor in neuroimaging beyond fMRIPrep},
+  author={Esteban, Oscar and Wright, Jessey and Markiewicz, Christopher J and Thompson, William Hedley and Goncalves, Mathias and Ciric, Rastko and Blair, Ross W and Feingold, Franklin and Rokem, Ariel and Ghosh, Satrajit and others},
+  year={2019},
+  publisher={OSF Preprints}
+}
+
+@misc{fitlins:2020,
+  author = {Markiewicz, Chriss and et al},
+  title = {FitLins - Fitting Linear Models to BIDS Datasets},
+  year = {2020},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/poldracklab/fitlins}},
+}
