+++ paper.bib
@@ -34,3 +34,14 @@
   journal = {GitHub repository},
   howpublished = {\url{https://github.com/poldracklab/fitlins}},
 }
+
+@article{abraham2014machine,
+  title={Machine learning for neuroimaging with scikit-learn},
+  author={Abraham, Alexandre and Pedregosa, Fabian and Eickenberg, Michael and Gervais, Philippe and Mueller, Andreas and Kossaifi, Jean and Gramfort, Alexandre and Thirion, Bertrand and Varoquaux, Ga{\"e}l},
+  journal={Frontiers in neuroinformatics},
+  volume={8},
+  pages={14},
+  year={2014},
+  publisher={Frontiers}
+}
+
