### 2021-02-01 vahid-sb/MiTfAT#1 @vahid-sb
Dear @emdupre. Thanks again for your suggestions. I decided to change the manual to refer only to the zip file rather than accompanying the data with the code. The sample datasets change the library size from KBs to tens of MBs, and I prefer to leave it to the user to download the dataset or not.
