### 2021-02-01 vahid-sb/MiTfAT#2 @vahid-sb
That filename should have been the name of the text file, which I changed in the manual. But that part is simply a very generic example. I have rearranged that chapter and made it (hopefully) easier to read and brought the discussions on the sample dataset to a new subsection to make things more organized.
