### 2021-02-01 fea207e
Create CONTRIBUTING.md
