+++ CONTRIBUTING.md
@@ -1,0 +1,28 @@
+# Contributing
+
+We love contributions from everyone. You can contribute following the follwing steps.  
+
+Fork, then clone the repo:
+
+    git clone git@github.com:your-username/MiTfAT.git
+
+Download the sample test scripts and dataset from [here].
+
+[here]: https://github.com/vahid-sb/MiTfAT/blob/master/tests.zip
+
+Makes sure the test scripts run without errors. 
+
+Make your change. Add tests for your change. 
+
+Push to your fork and [submit a pull request][pr].
+
+[pr]: https://github.com/vahid-sb/MiTfAT/compare/
+
+At this point you're waiting on us. We may come back to you with some changes or improvements or alternatives.
+
+Some things that will increase the chance that your pull request is accepted:
+
+* Write tests.
+* Write a [good commit message][commit].
+
+[commit]: http://tbaggery.com/2008/04/19/a-note-about-git-commit-messages.html
