### 2021-02-01 #2827 @vahid-sb
Dear @JonathanReardon. Thanks for your suggestions. About the issues on the paper text, I have made a number of changes and corrections to the file, including removing or correcting the issues you highlighted. And I added CONTRIBUTING.md addressing how people can contribute.
