### 2021-02-01 #2827 @arokem
Thanks @JonathanReardon.

@emdupre : have your concerns been properly addressed in the revision? Can https://github.com/vahid-sb/MiTfAT/issues/1, https://github.com/vahid-sb/MiTfAT/issues/2 and https://github.com/vahid-sb/MiTfAT/issues/3 be closed?
