### 2021-02-01 vahid-sb/MiTfAT#1 @emdupre
> I decided to change the manual to refer only to the zip file rather than accompanying the data with the code. The sample datasets change the library size from KBs to tens of MBs, and I prefer to leave it to the user to download the dataset or not.

That makes sense ! I'll follow-up on this in #2, but it might be worth highlighting the example data a bit more prominently in the installation instructions. Since the installation strictly works, though, I'll close this. Thanks for clarifying, here !
