### 2021-02-02 vahid-sb/MiTfAT#2 @emdupre
Thanks, @vahid-sb ! I'm able to run things much more smoothly now. I still am hitting a few stumbling blocks when following the docs that are worth addressing:

-  The first code blocks fails in section 3.1. I know that this is because it is only intended as an example, but I would argue that having the code blocks succeed is reasonably important ! I worry that the failure is likely to confuse users -- and you define an actual, usable filename just two sections down. I would recommend moving the `from mitfat.io import read_data` block into section 3.2 after you mention how to access the dataset.
- Chapter 4 onwards assumes we have a variable called `dataset1`, while Chapter 3 only defines`my_first_dataset`. This is small, but again will likely cause first user attempts to fail
