### 2021-02-02 #2827 @emdupre
> @emdupre : have your concerns been properly addressed in the revision? Can vahid-sb/MiTfAT#1, vahid-sb/MiTfAT#2 and vahid-sb/MiTfAT#3 be closed?

I've closed vahid-sb/MiTfAT#1 and vahid-sb/MiTfAT#3, but I still have two concerns that I've pointed to in vahid-sb/MiTfAT#2 ! If those are addressed then I'd be happy to recommend the package for publication.

Also, although i don't think this should hold up the software, I did want to point to two potential improvements re: package testing and contributing guidelines.

1. Currently, the testing is done manually (i.e, it must be run by a developer). I confirmed on my own machine that it passes, but it would be great to have this running on a continuous integration framework like CircleCI !
2. Also, the contributing guidelines are a bit sparse at present. If you're interested in expanding them, I would recommend pointing to or incorporating existing guidelines like [those from scikit-learn](https://scikit-learn.org/stable/developers/contributing.html) or [fMRIPprep](https://github.com/nipreps/fmriprep/blob/master/CONTRIBUTING.md).
