### 2021-02-07 vahid-sb/MiTfAT#2 @vahid-sb
Thanks for your diligence in checking such details in the manual. I changed Section 3.1 code to what is written in the sample script. So, it will load the sample dataset assuming the working directory is set to where the files are. Also, dataset1 is what is defined in Section 3.1 and clarified that in the future chapter, the variable refers to what is defined in Section 3.1.
