### 2021-02-07 #2827 @vahid-sb
@emdupre for CONTRIBUITING.md, I also checked a few of the samples in libraries I know. LIbraries like scikit-learn have a huge contributing community and such detailed instructions are not necessary. But given the fact that so far I have been the sole developer, did not want to put off a possible contributor with a long list of detailed instructions and kept it as simple as possible. And for CI, it has been on my mind and is in my long-term plans, but I am sure you agree that it is not something that undermines the functionalities of the library at this stage.
