### 2021-02-08 fce6337
Manual updated. Minor changes.
