### 2021-02-09 #2827 @emdupre
> But given the fact that so far I have been the sole developer, did not want to put the possible contributors off with a long list of detailed instructions and kept it as simple as possible. If the community around MiTfAT grows, then I'd update the contribution guidelines. And for CI, it has been on my mind and is in my long-term plans, but I am sure you agree that it is not something that undermines the functionalities of the library at this stage.

Yes, as I mentioned : I don't want to hold up the package over this, but I do think these are both directions that could be expanded in the future. Clearly that future is coming quickly -- I see that you've already gotten a [CI infrastructure](https://app.circleci.com/pipelines/github/vahid-sb/MiTfAT/3/workflows/49877be5-65d7-4bec-9e51-239116972cb2/jobs/3) set-up ! :slightly_smiling_face:

@arokem  : I've now closed https://github.com/vahid-sb/MiTfAT/issues/2, and I'm happy to recommend this software for publication. Thank you to @vahid-sb for addressing my comments and for putting this project together !
