### 2021-02-10 #2827 @arokem
I have also added some comments/suggestions on the text of the paper in https://github.com/vahid-sb/MiTfAT/pull/4
