### 2021-02-10 58f1f49
update to JOSS .tex .bib files
