+++ paper.bib
@@ -6,7 +6,8 @@
 	number={41},
 	pages={20666--20671},
 	year={2019},
-	publisher={National Acad Sciences}
+	publisher={National Acad Sciences},
+	doi={https://doi.org/10.1073/pnas.1908503116},
 }
 
 @article{kent:2019,
@@ -16,14 +17,16 @@
   volume={4},
   number={41},
   pages={1295},
-  year={2019}
+  year={2019},
+	doi={https://doi.org/10.21105/joss.01295}
 }
 
 @article{esteban:2019,
   title={NiPreps: enabling the division of labor in neuroimaging beyond fMRIPrep},
   author={Esteban, Oscar and Wright, Jessey and Markiewicz, Christopher J and Thompson, William Hedley and Goncalves, Mathias and Ciric, Rastko and Blair, Ross W and Feingold, Franklin and Rokem, Ariel and Ghosh, Satrajit and others},
   year={2019},
-  publisher={OSF Preprints}
+  publisher={OSF Preprints},
+	doi={10.31219/osf.io/ujxp6},
 }
 
 @misc{fitlins:2020,
@@ -33,6 +36,7 @@
   publisher = {GitHub},
   journal = {GitHub repository},
   howpublished = {\url{https://github.com/poldracklab/fitlins}},
+	doi={https://doi.org/10.5281/zenodo.1306215},
 }
 
 @article{abraham2014machine,
@@ -42,5 +46,6 @@
   volume={8},
   pages={14},
   year={2014},
-  publisher={Frontiers}
+  publisher={Frontiers},
+	doi={10.3389/fninf.2014.00014},
 }
