+++ paper.md
@@ -20,31 +20,35 @@ bibliography: paper.bib
 
 # Summary
 
-Functional Magnetic Resonance Imaging, fMRI, is a technique used in neuroscience to measure brain activity based on any signal that can be measured in an MRI scanner. Normally, fMRI is used to detect changes associated with blood flow, but it can also be used to detect changes in concentrations of molecules with different magnetic properties, which are directly injected into the brain of a subject.
+Functional Magnetic Resonance Imaging, fMRI, is a technique used in neuroscience to measure brain activity based on any signal that can be measured in an MRI scanner. Normally, fMRI is used to detect changes associated with blood flow, but it can also be used to detect changes in concentrations of molecules with different magnetic properties, that might be directly injected into the brain of a subject.
 
-Regardless of the signal that is measured in fMRI recording, from a computational point of view, fMRI recordings will result in a number of time-series. And then those time-series should be analyzed to find the answers to various questions of interest. The length of the time-series depends on the number of time-steps in which we have measured the signals, and their number depends on how-many voxels we have measured (a voxel is a 3-dimensional pixel or the unit of volume in which each fMRI signal is measured). The size of each of these voxels depends on the magnetic flux density of the MRI scanner, measured in Tesla (T). The higher the magnetic flux density, the smaller the voxels can be, and the higher is the spatial resolution of the measurement. Hence, we end up with one time-series for each of the voxels arranged in a 3-dimensional structure. One characteristic of the fMRI measurements is that while they can have a high spatial resolution, and while they allow us to measure brain activity not only in the cortex, but also in deeper regions of the brain, their temporal resolution is not normally high. And that can provide challenges for researchers who need to analyze the fMRI data.
+Regardless of the signal that is measured in fMRI recordings, from a computational point of view, fMRI recordings will result in a number of time-series. And then those time-series should be analyzed to find the answers to various questions of interest. The length of the time-series depends on the number of time-steps in which we have measured the signals, and their number depends on how-many voxels we have measured (a voxel is a 3-dimensional pixel or the unit of volume in which each fMRI signal is measured). The size of each of these voxels depends on the magnetic flux density of the MRI scanner, measured in Tesla (T). The higher the magnetic flux density, the smaller the voxels can be, and the higher is the spatial resolution of the measurements. Hence, we end up with one time-series for each of the voxels arranged in a 3-dimensional structure. One characteristic of the fMRI measurements is that while they can have a high spatial resolution, and while they allow us to measure brain activity not only in the cortex but also in deeper regions of the brain, their temporal resolution is not normally high. And that can provide challenges for researchers who need to analyze the fMRI data.
 
 # Statement of need
 
 `MiTfAT` is a scikit-learn-friendly Python library to analyze fMRI data, with a focus on molecular fMRI experiments. It was primarily developed for the study which is presented in [@savic:2019].
 
[…]
