### 2021-02-10 7db366b
update to JOSS .tex .bib files
