+++ paper.md
@@ -34,17 +34,17 @@ The basic principle behind `MiTfAT` is that it imports all the relevant data of 
 
 There are various functionalities available to analyze and visualize the data in a number of ways. They include:
 
-    - Clustering the time-series using K-means clustering. Clustering can be done based on raw or normalized values in all time-steps, or the mean value of each time-series or slope of a linear-regression passing through the time-series. If the time-span of the recording is divided into a number of segments, clustering can be done on each segment or the combination of the values in the segments. `MiTfAT` uses `scikit-learn` for all machine learning functionalities. Hence, K-means can be easily replaced with any other clustering algorithm implemented in `scikit-learn`.
+  - Clustering the time-series using K-means clustering. Clustering can be done based on raw or normalized values in all time-steps, or the mean value of each time-series or slope of a linear-regression passing through the time-series. If the time-span of the recording is divided into a number of segments, clustering can be done on each segment or the combination of the values in the segments. `MiTfAT` uses `scikit-learn` for all machine learning functionalities. Hence, K-means can be easily replaced with any other clustering algorithm implemented in `scikit-learn`.
 
-    - Removing voxels with a low signal to noise ratio. This is done using a clustering algorithm in two stages. In the first stage, the algorithm removes the time-series corresponding to voxels in which signal to noise ratio is not high enough. And in the second stage, the time-series corresponding to the remaining voxels are clustered to identify the distribution pattern of the contrast agent.
+  - Removing voxels with a low signal to noise ratio. This is done using a clustering algorithm in two stages. In the first stage, the algorithm removes the time-series corresponding to voxels in which signal to noise ratio is not high enough. And in the second stage, the time-series corresponding to the remaining voxels are clustered to identify the distribution pattern of the contrast agent.
 
-    - Detrending. We can remove the general trends in time-series to make the transient changes more visible. As an example, if we cannot wait long enough until the concentration of our agent reaches the steady-state level, then transient variations in the signal caused by changes in experimental conditions, which is what we are interested in, might be obscured by such trends in the signal. `MiTfAT` can detrend the time-series and give us a signal which looks like the one we might expect in a steady-state condition. And then, transient changes due to experimental design could be quantifiable.
+  - Detrending. We can remove the general trends in time-series to make the transient changes more visible. As an example, if we cannot wait long enough until the concentration of our agent reaches the steady-state level, then transient variations in the signal caused by changes in experimental conditions, which is what we are interested in, might be obscured by such trends in the signal. `MiTfAT` can detrend the time-series and give us a signal which looks like the one we might expect in a steady-state condition. And then, transient changes due to experimental design could be quantifiable.
 
-    - Interpolation of time-series under varying conditions. Assume we have an experiment in which the total recording time is divided into 3 segments. In the second segment, we record under a different condition compared to the first and last segment. This can be, as an example, the occlusion of an artery which changes calcium concentration in the brain when the agent we have injected into the brain binds with calcium. Obviously, we want to quantify how the signal has changed in the second segment. In order to do so, we should interpolate the time-series in segment 2 based on values of segments 1 and 3, and then compare it with the actual measurements. `MiTfAT` provides such a functionality out of the box.
[…]
