### 2021-02-10 d91b8ff
update to JOSS .tex .bib files
