### 2021-02-10 1973cac
update to JOSS .tex .bib files
