+++ paper.bib
@@ -7,7 +7,7 @@
 	pages={20666--20671},
 	year={2019},
 	publisher={National Acad Sciences},
-	doi={https://doi.org/10.1073/pnas.1908503116},
+	doi={10.1073/pnas.1908503116},
 }
 
 @article{kent:2019,
@@ -18,7 +18,7 @@
   number={41},
   pages={1295},
   year={2019},
-	doi={https://doi.org/10.21105/joss.01295}
+	doi={10.21105/joss.01295}
 }
 
 @article{esteban:2019,
@@ -36,7 +36,7 @@
   publisher = {GitHub},
   journal = {GitHub repository},
   howpublished = {\url{https://github.com/poldracklab/fitlins}},
-	doi={https://doi.org/10.5281/zenodo.1306215},
+	doi={10.5281/zenodo.1306215},
 }
 
 @article{abraham2014machine,
@@ -48,4 +48,15 @@
   year={2014},
   publisher={Frontiers},
 	doi={10.3389/fninf.2014.00014},
+}
+
+@article{scikit:2011,
+  title={Scikit-learn: Machine learning in Python},
+  author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others},
+  journal={the Journal of machine Learning research},
+  volume={12},
+  pages={2825--2830},
+  year={2011},
+  publisher={JMLR. org},
+	doi={10.5555/1953048.2078195},
 }
