+++ paper.md
@@ -26,7 +26,7 @@ Regardless of the signal that is measured in fMRI recordings, from a computation
 
 # Statement of need
 
-`MiTfAT` is a scikit-learn-friendly Python library to analyze fMRI data, with a focus on molecular fMRI experiments. It was primarily developed for the study which is presented in [@savic:2019].
+`MiTfAT` is a Python library to analyze fMRI data, with a focus on molecular fMRI experiments. It was primarily developed for the study which is presented in [@savic:2019].
 
 There are already a few Python packages that are used by researchers to pre-process the fMRI time-series and then analyze them, for example [@esteban:2019], and [@kent:2019], which focus on very specific points of the analysis workflow. Even a more comprehensive library such as NiLearn [@abraham2014machine], which includes various visualization functionalities and machine learning tools to analyze fMRI data, does not provide a ready-made framework to contain various information and measurements obtained in molecular fMRI experiments. Therefore, the `MiTfAT` library was developed. It can be used for general fMRI time-series analysis, but in particular for signals obtained from molecular fMRI studies, i.e., the cases in which we measure the changes in concentration of molecules that might have been directly injected into the brain. The `MiTfAT` library incorporates all the information and data related to an experiment into a Python class object called `fmri_dataset`. And various attributes of this class can be used to identify, analyze and visualize the data related to each experiment. Such datasets can include various MRI measurements of the same subject, for example, T1-weighted and FISP signals measured almost simultaneously. In addition, the dataset can include many trials in which the same set of stimuli is presented or applied to a subject repeatedly.
 
@@ -34,7 +34,7 @@ The basic principle behind `MiTfAT` is that it imports all the relevant data of 
 
 There are various functionalities available to analyze and visualize the data in a number of ways. They include:
 
-  - Clustering the time-series using K-means clustering. Clustering can be done based on raw or normalized values in all time-steps, or the mean value of each time-series or slope of a linear-regression passing through the time-series. If the time-span of the recording is divided into a number of segments, clustering can be done on each segment or the combination of the values in the segments. `MiTfAT` uses `scikit-learn` for all machine learning functionalities. Hence, K-means can be easily replaced with any other clustering algorithm implemented in `scikit-learn`.
+  - Clustering the time-series using K-means clustering. Clustering can be done based on raw or normalized values in all time-steps, or the mean value of each time-series or slope of a linear-regression passing through the time-series. If the time-span of the recording is divided into a number of segments, clustering can be done on each segment or the combination of the values in the segments. `MiTfAT` uses `scikit-learn` [@scikit:2011] for all machine learning functionalities. Hence, K-means can be easily replaced with any other clustering algorithm implemented in `scikit-learn`.
 
   - Removing voxels with a low signal to noise ratio. This is done using a clustering algorithm in two stages. In the first stage, the algorithm removes the time-series corresponding to voxels in which signal to noise ratio is not high enough. And in the second stage, the time-series corresponding to the remaining voxels are clustered to identify the distribution pattern of the contrast agent.
 
@@ -47,7 +47,7 @@ There are various functionalities available to analyze and visualize the data in
  - Visualization. The `fmri_dataset` class includes various visualization functions. They include functions that can plot the raw, normalized, averaged, or linear approximations of the time-series in each voxel independently or as part of an overall plot representing each layer in the fMRI mask. Also, various functions to plot centroids resulting from clustering algorithms and also plots to show how voxels corresponding to each cluster are placed in the overall mask. Such plots can be quite informative when we want to study the diffusion of a contrast agent in the brain.
 
 # Examples
[…]
