### 2021-02-10 e437e86
update to JOSS .tex .bib files
