### 2021-02-10 b0dcbb5
update to JOSS .tex .bib files
