+++ paper.md
@@ -20,7 +20,7 @@ bibliography: paper.bib
 
 # Summary
 
-Functional Magnetic Resonance Imaging, fMRI, is a technique used in neuroscience to measure brain activity based on any signal that can be measured in an MRI scanner. Normally, fMRI is used to detect changes associated with blood flow, but it can also be used to detect changes in concentrations of molecules with different magnetic properties, that might be directly injected into the brain of a subject.
+Functional Magnetic Resonance Imaging, fMRI, is a technique used in neuroscience to measure brain activity based on any signal that can be measured in an MRI scanner. Normally, fMRI is used to detect changes associated with blood flow, but it can also be used to detect changes in concentrations of molecules with different magnetic properties that might be directly injected into the brain of a subject.
 
 Regardless of the signal that is measured in fMRI recordings, from a computational point of view, fMRI recordings will result in a number of time-series. And then those time-series should be analyzed to find the answers to various questions of interest. The length of the time-series depends on the number of time-steps in which we have measured the signals, and their number depends on how-many voxels we have measured (a voxel is a 3-dimensional pixel or the unit of volume in which each fMRI signal is measured). The size of each of these voxels depends on the magnetic flux density of the MRI scanner, measured in Tesla (T). The higher the magnetic flux density, the smaller the voxels can be, and the higher is the spatial resolution of the measurements. Hence, we end up with one time-series for each of the voxels arranged in a 3-dimensional structure. One characteristic of the fMRI measurements is that while they can have a high spatial resolution, and while they allow us to measure brain activity not only in the cortex but also in deeper regions of the brain, their temporal resolution is not normally high. And that can provide challenges for researchers who need to analyze the fMRI data.
 
@@ -40,7 +40,7 @@ There are various functionalities available to analyze and visualize the data in
 
   - Detrending. We can remove the general trends in time-series to make the transient changes more visible. As an example, if we cannot wait long enough until the concentration of our agent reaches the steady-state level, then transient variations in the signal caused by changes in experimental conditions, which is what we are interested in, might be obscured by such trends in the signal. `MiTfAT` can detrend the time-series and give us a signal which looks like the one we might expect in a steady-state condition. And then, transient changes due to experimental design could be quantifiable.
 
-  - Interpolation of time-series under varying conditions. Assume we have an experiment in which the total recording time is divided into 3 segments. In the second segment, we record under a different condition compared to the first and last segment. This can be, as an example, the occlusion of an artery which changes calcium concentration in the brain when the agent we have injected into the brain binds with calcium. Obviously, we want to quantify how the signal has changed in the second segment. In order to do so, we should interpolate the time-series in segment 2 based on values of segments 1 and 3, and then compare it with the actual measurements. `MiTfAT` provides such a functionality out of the box.
+  - Interpolation of time-series under varying conditions. Assume we have an experiment in which the total recording time is divided into 3 segments. In the second segment, we record under a different condition than the first and third segments. This can be, as an example, the occlusion of an artery that changes calcium concentration in the brain when the contrast agent we have injected into the brain binds with calcium and we want to quantify how the signal has changed in the second segment. In order to do so, we should interpolate the time-series in segment 2 based on values of segments 1 and 3, and then compare it with the actual measurements during the second segment. `MiTfAT` provides such a functionality out of the box.
 
  - Averaging over many trials. If we repeat an experiment many times, then it is usually of interest to average the measurements over all the trials. This can be useful if each measurement is noisy and we want to attenuate the effects of noise by averaging. `MiTfAT` provides such functionality.
 
