### 2021-02-10 #2827 @vahid-sb
@emdupre, thanks for your kind words. And bout CI, just after I saw your comment I saw an ad for Circle CI on a completely unrelated website. Not a so subtle hint by the gods of the internet! I tried it and it gives an error while trying to install dependencies. I should have a look when I can to see how this can be fixed. And thanks again for all your constructive and useful comments.
