### 2021-02-10 #2827 @vahid-sb
@arokem Thanks for bringing the errors to my attention. I also fixed a few other minor ones mainly to make the text read better and clarify a couple of points that I felt can be explained in a more clear manner. And also added citation info for scikit-learn and added doi for all. And then it occurred to me that I have not mentioned anything about visualization functions of the library, so I added a few lines about that too. Please let me know if there are any remaining outstanding issues.
