### 2021-02-16 8465161
update to JOSS .tex .bib files
