### 2021-02-17 #2827 @arfon
> @openjournals/joss-eics : this is ready for your final review

Thanks @arokem. We ask that editors run `@whedon accept` as the handoff to the EiC on rotation. This adds a special label (and makes the final proofs) for review too.
