### 2021-02-21 #2827 @arfon
@JonathanReardon, @emdupre - many thanks for your reviews here and to @arokem for editing this submission. JOSS relies upon the volunteer efforts of folks like yourselves and we simply wouldn't be able to do this without you! ✨
