### 2021-02-21 #2827 @vahid-sb
I wish to sincerely thank you all for your contributions to this review process. Dear @arokem, thanks for accepting to be the editor and for your effort in organizing the review and for your own comments to improve the paper. Dear @emdupre and @JonathanReardon, thanks a lot for accepting to be a reviewer and for your very helpful comments which made this library and its documentation better and more accessible and useful for potential users. And dear @arfon, thanks for being the driving force behind JOSS. As somebody who has done quite a few reviews for various journals, I find this style of review the best and the most transparent and something all journals should aspire to achieve. Thanks again to you all.
